==> clusterfeature/05a-test-reconstruct.ipynb <==
# Test Reconstruct

import numpy as np
import pandas as pd
import joblib
import sys

sys.path.append("..")
from helpers import filename_for

## Load saved model

model = joblib.load("models/15m_30dp_600AMeV_500keV_4n_RandomForestClassifier.pkl")
features = ["T", "E", "Size", "EToF", "EnergyMoment", "TSpawn", "MaxEHit", "X", "Y", "Z"]

## Test loaded model with single cluster data

model.predict_proba(
    [[63.421486, 1.582491, 1.0, 621.388550, 8.881784e-16, 0.000000, 1.582491, -18.501257, -7.500000, 1522.5]]
)

## Test reconstruction: Sort by probability

This works quite well - the clusters with the highest probability are likely to the ones we're looking for.

Strategy: Take the multiplicity *n* from somewhere else and the *n* clusters with the highest probability

file = filename_for(15, 30, 600, 500, 4, "inclxx", 0, "clusterfeature.pkl")
clus = pd.read_pickle(file)
clus["pred_prim_proba"] = model.predict_proba(clus[features])[:, 1]

for i in range(10):
    display(clus[clus["i_event"] == i].sort_values(["pred_prim_proba"], ascending=[False])[:7])
