==> clusterfeature/05b-test-root-reconstruct.ipynb <==
# Test model use and reconstruction in ROOT

import os
import sys
import ROOT

sys.path.append("..")
from helpers import filename_for

def test_reco(distance, doubleplane, energy, erel, neutron, physics, subrun):
    inpfile = filename_for(distance, doubleplane, energy, erel, neutron, physics, subrun, "digi.root")
    parfile = filename_for(distance, doubleplane, energy, erel, neutron, physics, subrun, "para.root")
    outfile = filename_for(distance, doubleplane, energy, erel, neutron, physics, subrun, "test.root")

    ROOT.ROOT.EnableThreadSafety()
    ROOT.FairLogger.GetLogger().SetLogVerbosityLevel("LOW")
    ROOT.FairLogger.GetLogger().SetLogScreenLevel("DEBUG")

    # I/O
    run = ROOT.FairRunAna()
    run.SetSource(ROOT.FairFileSource(os.fspath(inpfile)))
    run.SetSink(ROOT.FairRootFileSink(os.fspath(outfile)))

    # Connect Runtime Database
    rtdb = run.GetRuntimeDb()
    pario = ROOT.FairParRootFileIo(False)
    pario.open(os.fspath(parfile))
    rtdb.setFirstInput(pario)
    rtdb.setOutput(pario)
    rtdb.saveOutput()

    run.AddTask(ROOT.R3BNeulandMultiplicityCheat())
    run.AddTask(ROOT.R3BNeulandNeutronsScikit("models/15m_30dp_600AMeV_500keV_4n_RandomForestClassifier.pkl"))
    
    run.Init()
    run.Run(0, 5)
    run.TerminateRun()

test_reco(15, 30, 600, 500, 4, "inclxx", 0)
