==> clusterfeature/09-automl.ipynb <==
# Super Simple Keras model on TriFeature for multiplicity determination

features = ["T", "E", "Size", "EToF", "EnergyMoment", "TSpawn", "MaxEHit", "X", "Y", "Z"]
label = "prim"
dp = 30
neutrons = [1, 2, 3, 4]

import sys
from collections import defaultdict

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import joblib
import sklearn.preprocessing
from sklearn.metrics import (
    balanced_accuracy_score,
    confusion_matrix,
    plot_confusion_matrix,
)

sys.path.append("..")
from helpers import filename_for

files = [filename_for(15, dp, 600, 500, n, "inclxx", s, "clusterfeature.parquet") for n in neutrons for s in range(20)]
dfs = [pd.read_parquet(file) for file in files]
data = pd.concat(dfs, ignore_index=True).sample(frac=1, random_state=1337)

data = data[(abs(data["X"]) < 125) & (abs(data["Y"]) < 125) & (data["T"] < 100)]

print(data["prim"].value_counts())
display(data)

#scaler = sklearn.preprocessing.RobustScaler()
#data[features] = scaler.fit_transform(data[features])
#joblib.dump(scaler, "models/keras-scaler.pkl.gz")
#display(data)

prim1 = data[data["prim"] == 1]
prim0 = data[data["prim"] == 0].sample(n=len(prim1.index), random_state=1337)
balanced_data = pd.concat([prim0, prim1], ignore_index=True).sample(frac=1, random_state=1337)

print(balanced_data["prim"].value_counts())

msk = np.random.rand(len(balanced_data)) < 0.8
traindata = balanced_data[msk]
testdata = balanced_data[~msk]

print(traindata.shape, testdata.shape)

x_train = traindata[features].values
y_train = traindata[[label]].values.ravel()

x_val = testdata[features].values
y_val = testdata[[label]].values.ravel()

#from autosklearn.experimental.askl2 import AutoSklearn2Classifier
from autosklearn.classification import AutoSklearnClassifier

cls = AutoSklearnClassifier(n_jobs=10, memory_limit=50000, time_left_for_this_task=36000)
cls.fit(x_train, y_train)

print(cls.sprint_statistics())

x_test = testdata[features].values
y_true = testdata[[label]].values.ravel()

y_pred = cls.predict(x_test)
bac = balanced_accuracy_score(y_true, y_pred)

print(bac)

#joblib.dump(cls, "models/autosklearn.pkl")
